Form Submission > Database Entry
Altered the controllers and models to utilize a templated view
(create_form.php) which is loaded based upon the tab.

I was trying to get the success.php view to load on a successful insert,
but failed to do so. Information submitted in the form will be inserted
into the proper database.