Loading items on homepage
Created item_view in the views folder. This uses the bh_list model to
query the database for some information. I have them loaded on the index
of each controller and query the database for cheap food/drink. They are
displayed in collapsable content forms, clicking on them reveals a
description.