Bar Template Page
Template page for every bar. Pulls information from locations table
(which I created) for the bar (name, address, description, rating). I
plan to use a JQuery plugin for the rating system but haven't done that
yet. There is a Google Maps API which is hardcoded for the moment but
should be configured using the address pulled from the DB.

The page displays the top drinks and top foods for that particular
location using location_id as a filter.

Here is the SQL to run for the new table:
--
-- Table structure for table `drink`
--

CREATE TABLE IF NOT EXISTS `drink` (
`drink_id` int(11) NOT NULL AUTO_INCREMENT COMMENT 'The primary key',
`name` varchar(100) NOT NULL COMMENT 'The human readable name of this
drink',
`description` text NOT NULL COMMENT 'An optional description about this
drink',
`price` decimal(6,0) NOT NULL COMMENT 'The price in dollars of a drink',
`location_id` int(11) NOT NULL COMMENT 'The id of the locaiton this
drink is at',
PRIMARY KEY (`drink_id`)
) ENGINE=InnoDB  DEFAULT CHARSET=latin1 COMMENT='This tables tracks data
relating to drinks. Each entry is a drink at a locaiton'
AUTO_INCREMENT=11 ;